source_code/requirements.txt
google-cloud-pubsub
google-cloud-bigquery
faker
